requirement.txt
torch>=2.0
torchvision
numpy
pandas
tqdm
lmdb
opencv-python
pillow
tensorboardx
wandb

